Some basic configuration.
This is the equivalent to writing a config/database.yml file in a Rails app, and
setting up a few other sensible defaults in config/application.rb

The only potentially surprising thing here is the addition of south. Django
doesn't ship with a migration system, but south is the defacto standard at this
point.

Now this is set up we can make a database, and run the migrations. The first
command has nothing to do with Django, it's a Postgres command, so you may need
to adjust it for your own situation:

    createdb django_blog
    python manage.py syncdb --migrate